requirements.txt
# RCP Unified — all dependencies
fastapi
uvicorn
anthropic
youtube-transcript-api
trafilatura
newspaper3k
python-dotenv
pydantic
httpx
feedparser
resend
python-multipart
tavily-python
